fix fetch calls and other bugs
