+++ src/routes/oyster/+layout.svelte
@@ -2,7 +2,10 @@
 	import NetworkPrompt from '$lib/components/prompts/NetworkPrompt.svelte';
 	import PageWrapper from '$lib/components/wrapper/PageWrapper.svelte';
 	import { getAllowance } from '$lib/controllers/contract/usdc';
-	import { getApprovedOysterAllowancesFromSubgraph } from '$lib/controllers/subgraphController';
+	import {
+		getAllProvidersDetailsFromSubgraph,
+		getApprovedOysterAllowancesFromSubgraph
+	} from '$lib/controllers/subgraphController';
 	import {
 		chainConfigStore,
 		chainStore,
@@ -13,9 +16,14 @@
 	import { environment } from '$lib/data-stores/environment';
 	import {
 		initializeAllowanceInOysterStore,
-		oysterTokenMetadataStore
+		oysterTokenMetadataStore,
+		oysterRateMetadataStore,
+		setMarketplaceLoadedInOysterStore,
+		updateMarketplaceDataInOysterStore
 	} from '$lib/data-stores/oysterStore';
 	import { connected, walletStore } from '$lib/data-stores/walletProviderStore';
+	import { getOysterProvidersModified } from '$lib/utils/data-modifiers/oysterModifiers';
+	import { addRegionNameToMarketplaceData } from '$lib/utils/helpers/oysterHelpers';
 	import type { BrowserProvider } from 'ethers';
 	import { onDestroy, onMount } from 'svelte';
 
@@ -23,7 +31,8 @@
 		setAllowedChainsStore(environment.supported_chains.oyster);
 	});
 
-	let prevChainId: null | number = null;
+	let prevChainIdCdata: null | number = null;
+	let prevChainIdMdata: null | number = null;
 	let prevAddress = '';
 
 	async function loadConnectedData() {
@@ -41,19 +50,37 @@
 		console.log('Oyster allowances data is loaded');
 	}
 
+	async function loadMarketplaceData() {
+		console.log('Loading marketplace data');
+		setMarketplaceLoadedInOysterStore(false);
+		const providersDataFromSubgraph = await getAllProvidersDetailsFromSubgraph();
+		const marketplaceData = await getOysterProvidersModified(
+			providersDataFromSubgraph,
+			$oysterRateMetadataStore.rateCPUrlUnitInSeconds
+		);
+		const marketplaceDataWithRegionName = addRegionNameToMarketplaceData(marketplaceData);
+		// updating stores instead of returning data as we don't need to show this data explicitly
+		updateMarketplaceDataInOysterStore(marketplaceDataWithRegionName);
+		console.log('marketplace data is loaded');
+	}
+
 	$: chainSupported = $chainStore.chainId
 		? $allowedChainsStore.includes($chainStore.chainId)
 		: true;
 
+	$: if ($chainStore.chainId !== prevChainIdMdata) {
+		prevChainIdMdata = $chainStore.chainId;
+		loadMarketplaceData();
+	}
+
 	$: if (
 		$connected &&
 		$walletStore.address !== prevAddress &&
-		$chainStore.chainId !== prevChainId &&
-		chainSupported
+		$chainStore.chainId !== prevChainIdCdata
 	) {
-		loadConnectedData();
-		prevChainId = $chainStore.chainId;
+		prevChainIdCdata = $chainStore.chainId;
 		prevAddress = $walletStore.address;
+		loadConnectedData();
 	}
 
 	onDestroy(() => {
